src/requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
asn1crypto==0.24.0
async-generator==1.10
attrs==20.3.0
Automat==0.6.0
backcall==0.2.0
bleach==3.2.1
blinker==1.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==6.7
cloud-init==20.2
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
entrypoints==0.3
et-xmlfile==1.0.1
Flask==0.12.2
httplib2==0.9.2
hyperlink==17.3.1
idna==2.10
incremental==16.10.1
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==0.24
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonpatch==1.16
jsonpointer==1.10
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mypy==0.560
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
netifaces==0.10.4
notebook==6.1.5
numpy==1.19.5
oauthlib==2.0.6
openpyxl==3.0.6
packaging==20.8
PAM==0.4.2
pandas==1.2.0
pandas-datareader==0.9.0
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
psutil==5.4.2
ptyprocess==0.6.0
py==1.10.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
Pygments==2.7.3
pygobject==3.26.1
pyinotify==0.9.6
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.4
pytest==6.2.1
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
pyzmq==20.0.0
requests==2.25.1
requests-unixsocket==0.1.5
scipy==1.6.0
seaborn==0.11.1
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==16.0.0
simplejson==3.13.2
six==1.15.0
sqlparse==0.4.1
ssh-import-id==5.7
systemd-python==234
terminado==0.9.1
testpath==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
Twisted==17.9.0
typed-ast==1.1.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
virtualenv==15.1.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.14.1
zope.interface==4.3.2
